pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xpertagent"
version = "0.1.1"
description = "XpertAgent - A flexible and powerful AI agent framework"
authors = [
    {name = "rookielittblack", email = "rookielittblack@yeah.net"},
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache License 2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache License 2.0",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=0.19.0",
    "httptools>=0.6.4",
    "latex2mathml>=3.77.0",
    "pipdeptree>=2.23.4",
    "socksio>=1.0.0",
    "uvloop>=0.21.0",
    "watchfiles>=0.24.0",
    "wavedrom>=2.0.3.post3",
    "websockets>=14.1",
    "wheel>=0.44.0",
    "pymongo>=4.10.1",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0"
]

[project.urls]
Homepage = "https://github.com/rookie-littleblack/XpertAgent"

[tool.setuptools]
packages = ["xpertagent"]
package-dir = {"GOT" = "xpertagent/tools/xpert_ocr/vendor/got_ocr/GOT-OCR-2.0-master"}

[tool.setuptools.package-data]
"*" = ["README.md"]